Fix "already initialized constant warnings" when using spring
To properly re(load) the constants and modules they should be located
in the path the auto(re)loader expects them.
